Ziegenhain et al. 2017: Gene Expression Matrices

J1 mESC were cultured in ground pluripotency states by 2i/LIF culture as described previously.
The cells were sorted (FACS) and processed in two batches.
Library preparations were performed with different protocols. Furthermore, spike-ins (ERCC) were added.
"CELseq2" : Library preparations were performed with CEL-seq2 on the Fluidigm C1, as described in Hashimshony et al. 2016.
"DropSeq": Drop-seq protocol by emulsion of single-cells with primer-microbreads, as described in Macosko et al. 2015.
"MARSseq" : MARS-seq protocol from FACS-sorted single cells as described in Jaitin et al. 2014.
"SCRBseq" : SCRB-seq protocol from FACS-sorted single cells as described in Soumillon et al. 2014.
"SmartSeq" : Library preparation was performed with the Fluidigm C1 platform as per manufacturers' protocol.
"SmartSeq2" : Library preparations were performed using the Smart-seq2 protocol, as described in Picelli et al. 2013.

Data Processing:
base-calls were performed with Illumina bcl2fastq.
All cDNA reads were aligned with Star v2.4.0.
SCRB-seq and Drop-seq protocol reads were tagged with its mate-pair UMI and cell barcode; unique reads were obtained.
Smart-seq and Smart-seq2 reads were assigned to genes using Rsubread.
Only cells were considered with at least 1,000,000 reads.
The R Data files are once for Read Counts and once for UMI Counts per Library Preparation Protocol.

data(CELseq2_Gene_Read_Counts)
data(CELseq2_Gene_UMI_Counts)
data(DropSeq_Gene_Read_Counts)
data(DropSeq_Gene_UMI_Counts)
data(MARSseq_Gene_Read_Counts)
data(MARSseq_Gene_UMI_Counts)
data(SCRBseq_Gene_Read_Counts)
data(SCRBseq_Gene_UMI_Counts)
data(SmartSeq_Gene_Read_Counts)
data(SmartSeq2_Gene_Read_Counts)

Format

An object of class "data.frame", with genes per row and cells per column.

Source

https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790.