Model Diagnostics for RNAseq Data

With this function, the user can determine goodness of fit for each gene.

evaluateDist(countData, batchData =NULL,
spikeData = NULL, spikeInfo = NULL,
Lengths = NULL, MeanFragLengths = NULL,
RNAseq, normalisation,
frac.genes=1, min.meancount = 0.1,
max.dropout=0.7, min.libsize=1000,
verbose = TRUE)

Arguments

countData
is a count matrix (row=gene, column=sample). Please provide the measurements of one group only, e.g. the control group.

batchData
is a data.frame for batch annotation. Rows correspond to samples. The first column should contain the batches, e.g. 'a', 'b', 'c', etc.. This is only used for the normalisation.

spikeData
is a count matrix. Rows correspond to spike-ins, columns to samples. The order of columns should be the same as in the countData.

spikeInfo
is a molecule count matrix of spike-ins. Rows correspond to spike-ins. The order of rows should be the same as in the spikeData. The column names should be 'SpikeID' and 'SpikeInput' for molecule counts of spike-ins.

Lengths
is a numeric vector of transcript lengths with the same length and order as the rows in countData. This variable is only used for internal TPM calculations if Census normalization is specified.

MeanFragLengths
is a numeric vector of mean fragment lengths with the same length as columns in countData. This variable is only used for internal TPM calculations if Census normalization is specified.

RNAseq
is a character value: "bulk" or "singlecell".

normalisation
is a character value: 'TMM', 'MR', 'PosCounts', 'UQ', 'scran', 'Linnorm', 'SCnorm', 'RUV', 'Census', 'depth', 'none'. For more information, please consult the details section of estimateParam.

frac.genes
The fraction of genes to calculate goodness of fit statistics, default is 1, i.e. for all genes.

min.meancount
The minimum mean normalised count per gene if fraction of genes are defined. Default is 0.1.

max.dropout
The maximal percentage of zero expression per gene. Genes with more than max.dropout will be remove. Default is 0.7, i.e. genes with more than 70% dropouts.

min.libsize
The minimum raw read counts per sample, default is 1000.

verbose
Logical value to indicate whether to show progress report of simulations.

Examples

# NOT RUN {
## using example data set
data(kolodziejczk_cnts)
evaldist <- evaluateDist(countData = kolodziejczk_cnts,
RNAseq = "singlecell", normalisation="scran",
frac.genes=1, min.meancount = 0.1,
max.dropout=0.7, min.libsize=1000,
verbose = TRUE)
plotEvalDist(evaldist, annot = TRUE)
# }