Receiver operator characteristics from simulation results

This function takes the simulation output from simulateDE and computes receiver operator characteristics (ROC) and area under the curve values (AUC).

evaluateROC(simRes,
alpha.type=c("adjusted","raw"),
MTC=c('BY', 'BH', 'Storey', 'IHW',
'holm', 'hochberg', 'hommel', 'bonferroni'),
alpha.nominal = 0.1,
target.by=c("lfc", "effectsize"),
delta=0)

Arguments

simRes
The result from simulateDE.

alpha.type
A string to represent the way to call DE genes. Available options are "adjusted" i.e. applying multiple testing correction and "raw" i.e. using p-values. Default is "adjusted".

MTC
Multiple testing correction method to use. Available options are 1) see p.adjust.methods, 2) Storey's qvalue see qvalue and 3) Independent Hypothesis Weighting considering mean expression as covariate (see ihw). Default is BY, i.e. Benjamini-Yekutieli FDR correction method.

alpha.nominal
The nomial value of significance. Default is 0.1.

target.by
A string to specify the method to define "biologically important" DE genes. Available options are (1) "lfc": interesting genes are defined by absolute log2 fold changes. (2) "effectsize": interesting genes are defined by absolute log2 fold changes divided by the square root of 1/(mean+dispersion).

delta
A threshold used for defining "biologically important" genes. Genes with absolute log2 fold changes (when target.by is "lfc") or effect sizes (when target.by is "effectsize") greater than this value are deemed DE in error rates calculations. If delta=0 then no threshold is applied. See "Details" for more description.

Value

A list with the following entries:

ROCR-Prediction

The output of prediction split up by group comparisons into lists.

ROCR-Performance

The output of performance calculating TPR and FPR for ROC curve. This is also split up by group comparisons into lists.

ROCR-AUC

The output of performance calculating AUC of ROC curve. split up by group comparisons into lists.

n1,n2

Sample sizes per group. This is taken from the simulation options.

target.by

The input method to define "biologically important" DE genes, either by log fold change or effect size.

delta

The input delta for biologically important genes. If delta=0, all target.by will be considered.

See also

Examples

# NOT RUN {
data(kolodziejczk_simDE)
eval.roc <- evaluateROC(simRes = kolodziejczk_simDE)
# }