Visualize simulated counts

This function performs multidimensional scaling of the simulated counts from the pairwise sample distances using variable genes (i.e. variance unequal to zero). Prior to distance calculation, the counts are normalized using the simulated size factors and log2 transformed. In the plot, the samples are annotated by phenotype and batch, if present.

plotCounts(simCounts, Distance, Scale, DimReduce, verbose = T)

Arguments

simCounts
The output of simulateCounts.

Distance
The (dis-)similarity measure to be used. This can be "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski" for distance measures and "spearman", "pearson" or "kendall" for correlation measures converted into distances, respectively. For more information, see dist and cor.

Scale
A logical vector indicating whether to use scaled log2 transformed counts or not.

DimReduce
The dimension reduction approach to be used. This can be "MDS" cmdscale, "PCA" prcomp, "t-SNE" Rtsne, "ICA" fastICA or "LDA" lda.

verbose
Logical value to indicate whether to print function information. Default is TRUE.

Value

A ggplot object.

Examples

# NOT RUN {
## not yet
# }