Visualize distributional characteristics of spike-ins

This function plots the results of the parameter estimation for spike-ins. This includes the absolute and relative sequencing depth (i.e. library size factor), a calibration curve as well as the capture efficiency given as a binomial regression.

plotSpike(estSpike, annot=TRUE)

Arguments

estSpike
The output of estimateParam.

annot
A logical vector. If TRUE, a short figure legend is included.

Value

A ggplot object.

Examples

# NOT RUN {
#' ## batch annotation
data(scrbseq_spike_cnts)
data(scrbseq_spike_info)
batch_info <- data.frame(Batch = ifelse(grepl(pattern = "SCRBseqA_",
colnames(scrbseq_spike_cnts)), "A", "B"),
row.names = colnames(scrbseq_spike_cnts))
## spike information table
spike_info <- scrbseq_spike_info[-1,]
## estimation
spike_param <- estimateSpike(spikeData = scrbseq_spike_cnts,
spikeInfo = spike_info,
MeanFragLength = NULL,
batchData = batch_info,
 = 'depth')
## plotting
plotSpike(estSpike = spike_param, annot=TRUE)
# }