Estimate simulation parameters for spike-ins

This function estimates and returns parameters needed for spike-in read count simulations using supplementary code from Kim et al. 2016 (DOI: 10.1038/ncomms9687).

estimateSpike(spikeData,
spikeInfo,
MeanFragLength = NULL,
batchData = NULL,
normalisation=c('depth','none'))

Arguments

spikeData
is a count matrix. Rows correspond to spike-ins, columns to samples. Rownames should contain the spike-in names, column names the sample names.

spikeInfo
is a molecule count matrix of spike-ins. Rows correspond to spike-ins. The order of rows should be the same as the rows in spikeData. The rownames should be the same as the rownames of spikeData. The first column should contain the molecule counts of spike-ins and be named 'SpikeInput'. The second column can contain the sequence lengths of spike-ins in bases and be named 'Lengths'.

MeanFragLength
is a numeric vector of the mean fragment length.

batchData
is a data.frame for batch annotation. Rows correspond to samples. The order of rows should be the same as the columns in spikeData. The rownames should be the same as the column names of spikeData. The first column should contain the batch annotation, e.g. 'a' for batch 1, 'b' for batch 2.

normalisation
is a character value: 'depth' or 'none'. For more information, please consult the details section.

Value

List object with the following entries:

normCounts

The normalised spike-in read counts data.frame.

normParams

The mean and standard deviation per spike-in using normalised read counts in a data.frame.

CaptureEfficiency

The ad-hoc estimated as well as fitted detection probabilities with confidence interval per spike-in using normalised read counts in a data.frame.

sf

The estimated library size factors.

seqDepth

Library size, i.e. total number of reads per library

seqDepth

The estimated library size factors.

EVGammaThetaEstimates

Estimation of the four parameters capturing technical variability, namely E[\(\gamma\)], Var[\(\gamma\)], E[\(\theta\)] and Var[\(\theta\)]. For more details, please consult supplementary information of Kim et al. 2016 (DOI: 10.1038/ncomms9687). These estimates are needed for simulating spike-in read counts.

Input

The input spike-in expression matrix, molecule counts data.frame and batch annotation data.frame, filtered so that only spike-ins with nonzero expression and samples with at least 100 reads are retained.

Settings

Reporting the chosen normalisation framework.

Details

Normalisation methods

'depth'

applies the depth normalization method as implemented in computeSpikeFactors.

'none'

No normalisation is applied. This approach can be used for prenormalized expression estimates, e.g. TPM/FPKM/RPKM estimated by cufflinks, RSEM, salmon etc.

Examples

# NOT RUN {
## batch annotation
data(scrbseq_spike_cnts)
data(scrbseq_spike_info)
batch_info <- data.frame(Batch = ifelse(grepl(pattern = "SCRBseqA_",
colnames(scrbseq_spike_cnts)), "A", "B"),
row.names = colnames(scrbseq_spike_cnts))
## spike information table
spike_info <- scrbseq_spike_info[-1,]
## estimation
spike_param <- estimateSpike(spikeData = scrbseq_spike_cnts,
spikeInfo = spike_info,
MeanFragLength = NULL,
batchData = batch_info,
normalisation = 'depth')
# }